order/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
aio-pika==9.5.0
psycopg2-binary==2.9.9
httpx==0.27.0
